refactor: update authentication module to use user repository
