--- repository/authentication/build.gradle.kts
@@ -1,53 +1,0 @@
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.serialization)
-    id("com.google.dagger.hilt.android")
-    alias(libs.plugins.ksp)
-}
-
-android {
-    namespace = "com.paris_2.repository.authentication"
-    compileSdk = Configurations.COMPILE_SDK
-
-    defaultConfig {
-        minSdk = Configurations.MIN_SDK_24
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-    kotlinOptions {
-        jvmTarget = Configurations.JVM_TARGET
-    }
-}
-
-dependencies {
-
-    implementation(libs.androidx.core.ktx)
-    implementation(libs.kotlinx.serialization.json)
-
-    //Hilt
-    implementation(libs.hilt.android)
-    ksp(libs.hilt.android.compiler)
-
-    // test
-    testImplementation(kotlin("test"))
-    testImplementation(libs.junit)
-    testImplementation(libs.mockk)
-    testImplementation(libs.truth)
-    testImplementation(libs.kotlinx.coroutines.test)
-
-    implementation(project(Modules.DOMAIN_USER))
-
-}
