--- repository/authentication/src/main/java/com/paris_2/repository/authentication/dataSource/local/AuthenticationLocalDataSource.kt
@@ -1,11 +1,0 @@
-package com.paris_2.repository.authentication.dataSource.local
-
-interface AuthenticationLocalDataSource {
-    fun saveSessionId(sessionId: String)
-    fun getSessionId(): String?
-    fun isLoggedIn(): Boolean
-
-    fun setIsGuest(isGuest: Boolean)
-    fun isGuest(): Boolean
-    fun hasAnySession(): Boolean
-}
