--- repository/authentication/src/main/java/com/paris_2/repository/authentication/exeptions/AuthenticationExceptions.kt
@@ -1,7 +1,0 @@
-package com.paris_2.repository.authentication.exeptions
-
-sealed class NetworkException(message: String? = null) : Exception(message) {
-    class UnauthenticatedException(message: String? = null) : NetworkException(message)
-    class ServerException(message: String? = null) : NetworkException(message)
-    class UnknownException(message: String? = null) : NetworkException(message)
-}
