--- repository/authentication/src/main/java/com/paris_2/repository/authentication/model/remote/SessionDto.kt
@@ -1,10 +1,0 @@
-package com.paris_2.repository.authentication.model.remote
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SessionDto(
-    @SerialName("session_id") val sessionId: String? = null,
-    @SerialName("success") val success: Boolean? = null
-)
