+++ repository/user/src/main/AndroidManifest.xml
