+++ repository/user/src/main/java/com/paris_2/repository/user/dataSource/local/AuthenticationLocalDataSource.kt
@@ -1,0 +1,11 @@
+package com.paris_2.repository.user.dataSource.local
+
+interface AuthenticationLocalDataSource {
+    fun saveSessionId(sessionId: String)
+    fun getSessionId(): String?
+    fun isLoggedIn(): Boolean
+
+    fun setIsGuest(isGuest: Boolean)
+    fun isGuest(): Boolean
+    fun hasAnySession(): Boolean
+}
