+++ repository/user/src/main/java/com/paris_2/repository/user/exeptions/AuthenticationExceptions.kt
@@ -1,0 +1,7 @@
+package com.paris_2.repository.user.exeptions
+
+sealed class NetworkException(message: String? = null) : Exception(message) {
+    class UnauthenticatedException(message: String? = null) : NetworkException(message)
+    class ServerException(message: String? = null) : NetworkException(message)
+    class UnknownException(message: String? = null) : NetworkException(message)
+}
