+++ repository/user/src/main/java/com/paris_2/repository/user/model/remote/RequestTokenDto.kt
@@ -1,0 +1,11 @@
+package com.paris_2.repository.user.model.remote
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestTokenDto(
+    @SerialName("expires_at") val expiresAt: String? = null,
+    @SerialName("request_token") val requestToken: String? = null,
+    @SerialName("success") val success: Boolean? = null
+)
