+++ settings.gradle.kts
@@ -84,5 +84,5 @@ include(":repository:home")
 include(":datasource:remote:home")
 include(":datasource:local:home")
 include(":datasource:remote:authentication")
-include(":repository:authentication")
+include(":repository:user")
 include(":datasource:local:authentication")
